refactor: 폴더 구조 변경 및 index로 파일 이름 적용
