+++ fe/src/Component/NewIssue/index.tsx
@@ -1,5 +1,5 @@
+import NewIssueOptions from "Component/IssueOptions";
 import NewIssueContent from "./NewIssueContent";
-import NewIssueOptions from "./NewIssueOptions";
 import { StyledNewIssue, StyledNewIssueHeader, StyledNewIssueMain } from "./NewIssue.styled";
 
 const WRITE_NEW_ISSUE = "WRITE NEW ISSUE";
